Caso de uso 07 Parte 1

+++ src/main/java/Facultad/TrabajoPracticoDesarrollo/DTOs/DtoResponsableDePago.java
@@ -22,5 +22,5 @@ import lombok.experimental.SuperBuilder;
         @JsonSubTypes.Type(value = DtoPersonaJuridica.class, name = "juridica")
 })
 public abstract class DtoResponsableDePago {
-    private Long idResponsable;
+    private Integer idResponsable;
 }
